Add Madeleine's ecology lectures and suggest instructors
Clearly not a completely fair distribution of lecturers right now, but it's a start for us to have discuss and have opinions on. I tried to adhere to the following:
- Lina wants to teach late in the course due to her committee meeting
- Madeleine is the one who knows most about the ecology topics
- Luke should teach his own package and the stats ecology class
- I should start since I will be doing introductions and be their point of contact

Currently each person has the following amount of lectures:
5 Madeleine
4 Joel
3 Luke
2 Lindsay
1.5 Lina and Elliot

A big reason for this is that later weeks have on lecture dedicated to project work, so there is only one teaching session. There might be a short teaching session in the beginning of the tutorial though. Week 11 might have a session added to it instead of the double project work (a result from packaging all the ecology lectures together)